refactor: better auth file structure
