+++ src/security/auth/authentication-modal/authentication-modal.scss
@@ -13,6 +13,7 @@
     }
 
     .container-bottom {
+      min-width: calc(5 * var(--var-size-15));
       display: flex;
       flex-direction: column;
       gap: var(--var-spacing-4);
