+++ src/security/login/login.html
@@ -1,0 +1,13 @@
+<form [formGroup]="loginForm">
+  <mat-form-field>
+    <mat-label>E-Mail</mat-label>
+    <input matInput formControlName="mail" />
+    <mat-error> {{'Leere oder ungültigte Eingabe'}} </mat-error>
+  </mat-form-field>
+  <mat-form-field>
+    <mat-label>Passwort</mat-label>
+    <input matInput formControlName="password" />
+    <mat-error> {{'Leere oder ungültigte Eingabe'}} </mat-error>
+  </mat-form-field>
+</form>
+<button (click)="login()" matButton>Login</button>
