--- src/security/register/register/register.html
@@ -1,20 +1,0 @@
-<div class="container-signup">
-  <form [formGroup]="signUpForm">
-    <mat-form-field>
-      <mat-label>E-Mail</mat-label>
-      <input matInput formControlName="mail" />
-      <mat-error> {{'Leere oder ungültigte Eingabe'}} </mat-error>
-    </mat-form-field>
-    <mat-form-field>
-      <mat-label>Passwort</mat-label>
-      <input matInput formControlName="password" />
-      <mat-error> {{'Leere oder ungültigte Eingabe'}} </mat-error>
-    </mat-form-field>
-    <mat-form-field>
-      <mat-label>Passwort Bestätigung</mat-label>
-      <input matInput formControlName="password_confirmation" />
-      <mat-error> {{'Leere oder ungültigte Eingabe'}} </mat-error>
-    </mat-form-field>
-  </form>
-  <button (click)="signup()" matButton>Registrieren</button>
-</div>
